Implement paged incremental queue updating
and fix more "same-name merged albums" instances
