+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "euphonica"
-version = "0.97.1"
+version = "0.97.2"
 edition = "2024"
 
 [dev-dependencies]
@@ -15,7 +15,7 @@ futures = "0.3.30"
 gettext-rs = { version = "0.7", features = ["gettext-system"] }
 image = "0.25.1"
 librsvg = "2.58.1"
-mpd = { git = "https://github.com/htkhiem/rust-mpd.git", rev = "6e5fdbe1352fe91df3097e3b1e5f05b40a8c41c5" }
+mpd = { git = "https://github.com/htkhiem/rust-mpd.git", rev = "04985377eb9c700d2237e6a895a689d204158f71" }
 once_cell = "1.19.0"
 time = { version = "0.3.36", features = ["formatting"] }
 reqwest = { version = "0.12.5", features = ["blocking", "json"] }
