+++ src/client/wrapper.rs
@@ -1,27 +1,24 @@
-use async_channel::{Receiver, SendError, Sender};
+use async_channel::{Receiver, Sender};
 use futures::executor;
 use glib::clone;
 use gtk::{gio, glib};
 use gtk::{gio::prelude::*, glib::BoxedAnyObject};
-use mpd::song::PosIdChange;
 use resolve_path::PathResolveExt;
 use mpd::error::ServerError;
 use mpd::{
     client::Client,
     error::{Error as MpdError, ErrorCode as MpdErrorCode},
     lsinfo::LsInfoEntry,
-    search::{Operation as QueryOperation, Query, Term, Window},
+    search::{Query, Term},
     song::Id,
     Channel, EditAction, Idle, Output, SaveMode, Subsystem,
 };
-use rustc_hash::FxHashSet;
 
 use std::net::TcpStream;
 use std::os::unix::net::UnixStream;
 use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::Arc;
 use std::{
-    borrow::Cow,
     cell::{Cell, RefCell},
     rc::Rc
 };
@@ -29,63 +26,19 @@ use uuid::Uuid;
 
 use crate::common::Stickers;
 use crate::{
-    common::{Album, AlbumInfo, Artist, ArtistInfo, INode, Song, SongInfo},
+    common::{Album, AlbumInfo, Artist, INode, Song, SongInfo},
     meta_providers::ProviderMessage,
-    cache::get_new_image_paths,
     player::PlaybackFlow,
     utils,
 };
 
+use super::{AsyncClientMessage, BackgroundTask};
 use super::state::{ClientState, ConnectionState, StickersSupportLevel};
 use super::stream::StreamWrapper;
 use super::password::get_mpd_password;
+use super::background;
 use super::ClientError;
 
-const BATCH_SIZE: usize = 1024;
-const FETCH_LIMIT: usize = 10000000; // Fetch at most ten million songs at once (same
-// folder, same tag, etc)
-
-// Messages to be sent from child thread or synchronous methods
-enum AsyncClientMessage {
-    Connect, // Host and port are always read from gsettings
-    Disconnect,
-    Status(usize), // Number of pending background tasks
-    Idle(Vec<Subsystem>), // Will only be sent from the child thread
-    QueueSongsDownloaded(Vec<SongInfo>),
-    QueueChangesReceived(Vec<PosIdChange>),
-    AlbumBasicInfoDownloaded(AlbumInfo), // Return new album to be added to the list model (as SongInfo of a random song in it).
-    RecentAlbumDownloaded(AlbumInfo),
-    AlbumSongInfoDownloaded(String, Vec<SongInfo>), // Return songs in the album with the given tag (batched)
-    ArtistBasicInfoDownloaded(ArtistInfo), // Return new artist to be added to the list model.
-    RecentArtistDownloaded(ArtistInfo),
-    ArtistSongInfoDownloaded(String, Vec<SongInfo>), // Return songs of an artist (or had their participation)
-    ArtistAlbumBasicInfoDownloaded(String, AlbumInfo), // Return albums that had this artist in their AlbumArtist tag.
-    FolderContentsDownloaded(String, Vec<LsInfoEntry>),
-    PlaylistSongInfoDownloaded(String, Vec<SongInfo>),
-    RecentSongInfoDownloaded(Vec<SongInfo>),
-    DBUpdated
-}
-
-// Work requests for sending to the child thread.
-// Completed results will be reported back via AsyncClientMessage.
-#[derive(Debug)]
-pub enum BackgroundTask {
-    Update,
-    DownloadFolderCover(AlbumInfo),
-    DownloadEmbeddedCover(SongInfo),
-    FetchQueue,  // Full fetch
-    FetchQueueChanges(u32),  // From given version
-    FetchFolderContents(String), // Gradually get all inodes in folder at path
-    FetchAlbums, // Gradually get all albums
-    FetchRecentAlbums,
-    FetchAlbumSongs(String),  // Get songs of album with given tag
-    FetchArtists(bool), // Gradually get all artists. If bool flag is true, will parse AlbumArtist tag
-    FetchRecentArtists,
-    FetchArtistSongs(String), // Get all songs of an artist with given name
-    FetchArtistAlbums(String), // Get all albums of an artist with given name
-    FetchPlaylistSongs(String), // Get songs of playlist with given name
-    FetchRecentSongs(u32), // Get last n songs
-}
 // Thin wrapper around the blocking mpd::Client. It contains two separate client
 // objects connected to the same address. One lives on the main thread along
 // with the GUI and takes care of sending user commands to the daemon, while the
@@ -136,668 +89,6 @@ pub enum BackgroundTask {
 // The child thread will only be able to do so after finishing idling, but
 // incidentally, disconnecting the main thread's client will send an idle message,
 // unblocking the child thread and allowing it to check the flag.
-
-mod background {
-    use std::ops::Range;
-
-    use gtk::gdk;
-    use time::OffsetDateTime;
-
-    use crate::{cache::sqlite, utils::strip_filename_linux};
-
-    use super::*;
-    pub fn update_mpd_database(
-        client: &mut mpd::Client<StreamWrapper>,
-        sender_to_fg: &Sender<AsyncClientMessage>,
-    ) {
-        match client.update() {
-            Ok(_) => {
-                let _ = sender_to_fg.send_blocking(AsyncClientMessage::DBUpdated);
-            }
-            Err(MpdError::Io(_)) => {
-                // Connection error => attempt to reconnect
-                let _ = sender_to_fg.send_blocking(AsyncClientMessage::Connect);
-            }
-            Err(_) => {}
-        }
-    }
-
-    pub fn get_current_queue(
-        client: &mut mpd::Client<StreamWrapper>,
-        sender_to_fg: &Sender<AsyncClientMessage>
-    ) {
-        // TODO: batched reads to avoid blocking MPD server
-        // For now we're fetching the entire thing at once in the background, then
-        // sending batches of it to the UI to allow the UI thread some slack.
-        match client.queue() {
-            Ok(mut queue) => {
-                let mut idx: usize = 0;
-                let len = queue.len();
-                while idx < len {
-                    let end = std::cmp::min(idx + BATCH_SIZE, len);
-                    let _ = sender_to_fg.send_blocking(AsyncClientMessage::QueueSongsDownloaded(
-                        queue[idx..end]
-                            .iter_mut()
-                            .map(|mpd_song| SongInfo::from(std::mem::take(mpd_song)))
-                            .collect()
-                    ));
-                    idx += BATCH_SIZE;
-                }
-            }
-            Err(MpdError::Io(_)) => {
-                // Connection error => attempt to reconnect
-                let _ = sender_to_fg.send_blocking(AsyncClientMessage::Connect);
-            }
-            Err(_) => {}
-        }
-    }
-
-    pub fn get_queue_changes(
-        client: &mut mpd::Client<StreamWrapper>,
-        sender_to_fg: &Sender<AsyncClientMessage>,
-        curr_version: u32
-    ) {
-        // TODO: batched reads to avoid blocking MPD server
-        // For now we're fetching the entire thing at once in the background, then
-        // sending batches of it to the UI to allow the UI thread some slack.
-        match client.changesposid(curr_version) {
-            Ok(changes) => {
-                let mut idx: usize = 0;
-                let len = changes.len();
-                while idx < len {
-                    let end = std::cmp::min(idx + BATCH_SIZE, len);
-                    let _ = sender_to_fg.send_blocking(AsyncClientMessage::QueueChangesReceived(
-                        changes[idx..end].to_owned()
-                    ));
-                    idx += BATCH_SIZE;
-                }
-            }
-            Err(MpdError::Io(_)) => {
-                // Connection error => attempt to reconnect
-                let _ = sender_to_fg.send_blocking(AsyncClientMessage::Connect);
-            }
-            Err(_) => {}
-        }
-    }
-
-    fn download_embedded_cover_inner(
-        client: &mut mpd::Client<StreamWrapper>,
-        uri: String
-    ) -> Option<(gdk::Texture, gdk::Texture)> {
-        if let Some(dyn_img) = client
-            .readpicture(&uri)
-            .map_or(None, |bytes| utils::read_image_from_bytes(bytes))
-        {
-            let (hires, thumb) = utils::resize_convert_image(dyn_img);
-            let (path, thumbnail_path) = get_new_image_paths();
-            hires.save(&path)
-                 .expect(&format!("Couldn't save downloaded cover to {:?}", &path));
-            thumb.save(&thumbnail_path)
-                 .expect(&format!("Couldn't save downloaded thumbnail cover to {:?}", &thumbnail_path));
-            sqlite::register_cover_key(
-                &uri, Some(path.file_name().unwrap().to_str().unwrap()), false
-            ).join().unwrap().expect("Sqlite DB error");
-            sqlite::register_cover_key(
-                &uri, Some(thumbnail_path.file_name().unwrap().to_str().unwrap()), true
-            ).join().unwrap().expect("Sqlite DB error");
-            let hires_tex = gdk::Texture::from_filename(&path).unwrap();
-            let thumb_tex = gdk::Texture::from_filename(&thumbnail_path).unwrap();
-            Some((hires_tex, thumb_tex))
-        } else {
-            None
-        }
-    }
-
-    fn download_folder_cover_inner(
-        client: &mut mpd::Client<StreamWrapper>,
-        folder_uri: String
-    ) -> Option<(gdk::Texture, gdk::Texture)> {
-        if let Some(dyn_img) = client
-            .albumart(&folder_uri)
-            .map_or(None, |bytes| utils::read_image_from_bytes(bytes))
-        {
-            let (hires, thumb) = utils::resize_convert_image(dyn_img);
-            let (path, thumbnail_path) = get_new_image_paths();
-            hires.save(&path)
-                 .expect(&format!("Couldn't save downloaded cover to {:?}", &path));
-            thumb.save(&thumbnail_path)
-                 .expect(&format!("Couldn't save downloaded thumbnail cover to {:?}", &thumbnail_path));
-            sqlite::register_cover_key(
-                &folder_uri, Some(path.file_name().unwrap().to_str().unwrap()), false
-            ).join().unwrap().expect("Sqlite DB error");
-            sqlite::register_cover_key(
-                &folder_uri, Some(thumbnail_path.file_name().unwrap().to_str().unwrap()), true
-            ).join().unwrap().expect("Sqlite DB error");
-            let hires_tex = gdk::Texture::from_filename(&path).unwrap();
-            let thumb_tex = gdk::Texture::from_filename(&thumbnail_path).unwrap();
-            Some((hires_tex, thumb_tex))
-        } else {
-            None
-        }
-    }
-
-    pub fn download_embedded_cover(
-        client: &mut mpd::Client<StreamWrapper>,
-        sender_to_cache: &Sender<ProviderMessage>,
-        key: SongInfo
-    ) {
-        // Still prioritise folder-level art if allowed to
-        let folder_uri = strip_filename_linux(&key.uri).to_owned();
-        // Re-check in case previous iterations have already downloaded these.
-        // Check using thumbnail = true to quickly refresh cache after a deletion of the entire
-        // images folder. This is because upon startup we'll mass-schedule thumbnail fetches, so
-        // in case the folder has been deleted, only thumbnail records in the SQLite DB will be
-        // dropped. Checking with thumbnail=true will still return a path even though that
-        // path has already been deleted, preventing downloading from proceeding.
-        let folder_path = sqlite::find_cover_by_key(&folder_uri, true).expect("Sqlite DB error");
-        if folder_path.is_none() {
-            if let Some((hires_tex, thumb_tex)) = download_folder_cover_inner(client, folder_uri.clone()) {
-                sender_to_cache
-                    .send_blocking(ProviderMessage::CoverAvailable(folder_uri.clone(), false, hires_tex))
-                    .expect("Cannot notify main cache of folder cover download result.");
-                sender_to_cache
-                    .send_blocking(ProviderMessage::CoverAvailable(folder_uri, true, thumb_tex))
-                    .expect("Cannot notify main cache of folder cover download result.");
-                return;
-            } // No folder-level art was available. Proceed to actually fetch embedded art.
-        } else if folder_path.as_ref().map_or(false, |p| p.len() > 0) {
-            // Nothing to do, as there's already a path in the DB.
-            return;
-        }
-        // Re-check in case previous iterations have already downloaded these.
-        let uri = key.uri.to_owned();
-        if sqlite::find_cover_by_key(&uri, true).expect("Sqlite DB error").is_none() {
-            if let Some((hires_tex, thumb_tex)) = download_embedded_cover_inner(client, uri.clone()) {
-                sender_to_cache
-                    .send_blocking(ProviderMessage::CoverAvailable(uri.clone(), false, hires_tex))
-                    .expect("Cannot notify main cache of embedded cover download result.");
-                sender_to_cache
-                    .send_blocking(ProviderMessage::CoverAvailable(uri, true, thumb_tex))
-                    .expect("Cannot notify main cache of embedded cover download result.");
-                return;
-            }
-            if let Some(album) = &key.album {
-                // Go straight to external metadata providers since we've already
-                // failed to fetch folder-level cover from MPD at this point.
-                // Don't schedule again if we've come back empty-handed once before.
-                if folder_path.is_none() {
-                    sender_to_cache
-                        .send_blocking(ProviderMessage::FetchFolderCoverExternally(album.clone()))
-                        .expect("Cannot signal main cache to run fallback folder cover logic.");
-                    return;
-                }
-            }
-            sender_to_cache
-                .send_blocking(ProviderMessage::CoverNotAvailable(uri))
-                .expect("Cannot notify main cache of embedded cover download result.");
-        } else {
-            // Nothing to do, as there's already a path in the DB
-            return;
-        }
-    }
-
-    pub fn download_folder_cover(
-        client: &mut mpd::Client<StreamWrapper>,
-        sender_to_cache: &Sender<ProviderMessage>,
-        key: AlbumInfo
-    ) {
-        // Re-check in case previous iterations have already downloaded these.
-        if sqlite::find_cover_by_key(&key.folder_uri, true).expect("Sqlite DB error").is_none() {
-            let folder_uri = key.folder_uri.to_owned();
-            if let Some((hires_tex, thumb_tex)) = download_folder_cover_inner(client, folder_uri.clone()) {
-                sender_to_cache
-                    .send_blocking(ProviderMessage::CoverAvailable(key.folder_uri.clone(), false, hires_tex))
-                    .expect("Cannot notify main cache of folder cover download result.");
-                sender_to_cache
-                    .send_blocking(ProviderMessage::CoverAvailable(key.folder_uri, true, thumb_tex))
-                    .expect("Cannot notify main cache of folder cover download result.");
-            } else {
-                // Fall back to embedded art.
-                let uri = key.example_uri.to_owned();
-                let sqlite_path = sqlite::find_cover_by_key(&uri, true).expect("Sqlite DB error");
-                if sqlite_path.is_none() {
-                    if let Some((hires_tex, thumb_tex)) = download_embedded_cover_inner(client, uri.clone()) {
-                        sender_to_cache
-                            .send_blocking(ProviderMessage::CoverAvailable(uri.clone(), false, hires_tex))
-                            .expect("Cannot notify main cache of embedded fallback download result.");
-                        sender_to_cache
-                            .send_blocking(ProviderMessage::CoverAvailable(uri, true, thumb_tex))
-                            .expect("Cannot notify main cache of embedded fallback download result.");
-                        return;
-                    }
-                } else if sqlite_path.as_ref().map_or(false, |p| p.len() > 0) {
-                    // Nothing to do, as there's already a path in the DB.
-                    return;
-                }
-                sender_to_cache
-                    .send_blocking(ProviderMessage::FetchFolderCoverExternally(key))
-                    .expect("Cannot signal main cache to fetch cover externally.");
-            }
-        }
-    }
-
-    // Err is true when a reconnection should be attempted
-    fn fetch_albums_by_query<F>(client: &mut mpd::Client<StreamWrapper>, query: &Query, respond: F) -> Result<(), bool>
-    where
-        F: Fn(AlbumInfo) -> Result<(), SendError<AsyncClientMessage>>,
-    {
-        // TODO: batched windowed retrieval
-        // Get list of unique album tags, grouped by albumartist
-        // Will block child thread until info for all albums have been retrieved.
-        match client.list(&Term::Tag(Cow::Borrowed("album")), query, Some("albumartist")) {
-            Ok(grouped_vals) => {
-                for (key, tags) in grouped_vals.groups.into_iter() {
-                    for tag in tags.iter() {
-                        match client.find(
-                            Query::new().and(Term::Tag(Cow::Borrowed("album")), tag).and(Term::Tag(Cow::Borrowed("albumartist")), &key),
-                            Window::from((0, 1)),
-                        ) {
-                            Ok(mut songs) => {
-                                if !songs.is_empty() {
-                                    let info = SongInfo::from(std::mem::take(&mut songs[0]))
-                                        .into_album_info()
-                                        .unwrap_or_default();
-                                    let _ = respond(info);
-                                }
-                            }
-                            Err(e) => {
-                                dbg!(e);
-                            }
-                        }
-                    }
-                }
-                Ok(())
-            }
-            Err(MpdError::Io(_)) => {
-                // Connection error => attempt to reconnect
-                Err(true)
-            }
-            Err(e) => {
-                dbg!(e);
-                Err(false)
-            }
-        }
-    }
-
-    fn fetch_songs_by_query<F>(client: &mut mpd::Client<StreamWrapper>, query: &Query, respond: F) -> Result<(), bool>
-    where
-        F: Fn(Vec<SongInfo>) -> Result<(), SendError<AsyncClientMessage>>,
-    {
-        let mut curr_len: usize = 0;
-        let mut more: bool = true;
-        while more && (curr_len) < FETCH_LIMIT {
-            match client
-                .find(query, Window::from((curr_len as u32, (curr_len + BATCH_SIZE) as u32))) {
-                    Ok(mut mpd_songs) => {
-                        let songs: Vec<SongInfo> = mpd_songs
-                            .iter_mut()
-                            .map(|mpd_song| SongInfo::from(std::mem::take(mpd_song)))
-                            .collect();
-                        if !songs.is_empty() {
-                            let _ = respond(songs);
-                            curr_len += BATCH_SIZE;
-                        } else {
-                            more = false;
-                        }
-                    }
-                    Err(MpdError::Io(_)) => {
-                        // Connection error => attempt to reconnect
-                        return Err(true);
-                    }
-                    Err(_) => {
-                        return Err(false);
-                    }
-                }
-        }
-        Ok(())
-    }
-
-    /// Fetch all albums, using AlbumArtist to further disambiguate same-named ones.
-    pub fn fetch_all_albums(client: &mut mpd::Client<StreamWrapper>, sender_to_fg: &Sender<AsyncClientMessage>) {
-        match fetch_albums_by_query(client, &Query::new(), |info| {
-            sender_to_fg.send_blocking(AsyncClientMessage::AlbumBasicInfoDownloaded(info))
-        }) {
-            Err(true) => {
-                let _ = sender_to_fg.send_blocking(AsyncClientMessage::Connect);
-            }
-            _ => {}
-        }
-    }
-
-    pub fn fetch_recent_albums(
-        client: &mut mpd::Client<StreamWrapper>,
-        sender_to_fg: &Sender<AsyncClientMessage>
-    ) {
-        let settings = utils::settings_manager().child("library");
-        let recent_albums = sqlite::get_last_n_albums(settings.uint("n-recent-albums")).expect("Sqlite DB error");
-        for tup in recent_albums.into_iter() {
-            let mut query = Query::new();
-            query.and(Term::Tag(Cow::Borrowed("album")), tup.0);
-            if let Some(artist) = tup.1 {
-                query.and(Term::Tag(Cow::Borrowed("albumartist")), artist);
-            }
-            if let Some(mbid) = tup.2 {
-                query.and(Term::Tag(Cow::Borrowed("musicbrainz_albumid")), mbid);
-            }
-            match fetch_albums_by_query(
-                client,
-                &query,
-                |info| {
-                    sender_to_fg.send_blocking(AsyncClientMessage::RecentAlbumDownloaded(info))
-                }) {
-                Err(true) => {
-                    let _ = sender_to_fg.send_blocking(AsyncClientMessage::Connect);
-                }
-                _ => {}
-            }
-        }
-    }
-
-    pub fn fetch_albums_of_artist(
-        client: &mut mpd::Client<StreamWrapper>,
-        sender_to_fg: &Sender<AsyncClientMessage>,
-        artist_name: String,
-    ) {
-        match fetch_albums_by_query(
-            client,
-            Query::new().and_with_op(
-                Term::Tag(Cow::Borrowed("artist")),
-                QueryOperation::Contains,
-                artist_name.clone(),
-            ),
-            |info| {
-                sender_to_fg.send_blocking(AsyncClientMessage::ArtistAlbumBasicInfoDownloaded(
-                    artist_name.clone(),
-                    info,
-                ))
-            },
-        ) {
-            Err(true) => {
-                let _ = sender_to_fg.send_blocking(AsyncClientMessage::Connect);
-            }
-            _ => {}
-        }
-    }
-
-    pub fn fetch_album_songs(
-        client: &mut mpd::Client<StreamWrapper>,
-        sender_to_fg: &Sender<AsyncClientMessage>,
-        tag: String,
-    ) {
-        match fetch_songs_by_query(
-            client,
-            Query::new().and(Term::Tag(Cow::Borrowed("album")), tag.clone()),
-            |songs| {
-                sender_to_fg.send_blocking(AsyncClientMessage::AlbumSongInfoDownloaded(
-                    tag.clone(),
-                    songs,
-                ))
-            },
-        ) {
-            Err(true) => {
-                let _ = sender_to_fg.send_blocking(AsyncClientMessage::Connect);
-            }
-            _ => {}
-        }
-    }
-
-    pub fn fetch_artists(
-        client: &mut mpd::Client<StreamWrapper>,
-        sender_to_fg: &Sender<AsyncClientMessage>,
-        use_album_artist: bool,
-    ) {
-        // Fetching artists is a bit more involved: artist tags usually contain multiple artists.
-        // For the same reason, one artist can appear in multiple tags.
-        // Here we'll reuse the artist parsing code in our SongInfo struct and put parsed
-        // ArtistInfos in a Set to deduplicate them.
-        let tag_type: &'static str = if use_album_artist {
-            "albumartist"
-        } else {
-            "artist"
-        };
-        let mut already_parsed: FxHashSet<String> = FxHashSet::default();
-        match client.list(&Term::Tag(Cow::Borrowed(tag_type)), &Query::new(), None) {
-            Ok(grouped_vals) => {
-                // TODO: Limit tags to only what we need locally
-                for tag in &grouped_vals.groups[0].1 {
-                    if let Ok(mut songs) = client.find(
-                        Query::new().and(Term::Tag(Cow::Borrowed(tag_type)), tag),
-                        Window::from((0, 1)),
-                    ) {
-                        if !songs.is_empty() {
-                            let first_song = SongInfo::from(std::mem::take(&mut songs[0]));
-                            let artists = first_song.into_artist_infos();
-                            // println!("Got these artists: {artists:?}");
-                            for artist in artists.into_iter() {
-                                if already_parsed.insert(artist.name.clone()) {
-                                    // println!("Never seen {artist:?} before, inserting...");
-                                    let _ = sender_to_fg.send_blocking(
-                                        AsyncClientMessage::ArtistBasicInfoDownloaded(artist),
-                                    );
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-            Err(MpdError::Io(_)) => {
-                // Connection error => attempt to reconnect
-                let _ = sender_to_fg.send_blocking(AsyncClientMessage::Connect);
-            }
-            _ => {}
-        }
-    }
-
-    pub fn fetch_recent_artists(
-        client: &mut mpd::Client<StreamWrapper>,
-        sender_to_fg: &Sender<AsyncClientMessage>
-    ) {
-        let mut already_parsed: FxHashSet<String> = FxHashSet::default();
-        let settings = utils::settings_manager().child("library");
-        let n = settings.uint("n-recent-artists");
-        let mut res: Vec<ArtistInfo> = Vec::with_capacity(n as usize);
-        let recent_names = sqlite::get_last_n_artists(n).expect("Sqlite DB error");
-        let mut recent_names_set: FxHashSet<String> = FxHashSet::default();
-        for name in recent_names.iter() {
-            recent_names_set.insert(name.clone());
-        }
-        for name in recent_names.iter() {
-            match client.find(
-                Query::new().and_with_op(Term::Tag(Cow::Borrowed("artist")), QueryOperation::Contains, name),
-                Window::from((0, 1))
-            ) {
-                Ok(mut songs) => {
-                    if !songs.is_empty() {
-                        let first_song = SongInfo::from(std::mem::take(&mut songs[0]));
-                        let artists = first_song.into_artist_infos();
-                        for artist in artists.into_iter() {
-                            if recent_names_set.contains(&artist.name) {
-                                if already_parsed.insert(artist.name.clone()) {
-                                    res.push(artist);
-                                }
-                            }
-                        }
-                    }
-                }
-                Err(MpdError::Io(_)) => {
-                    // Connection error => attempt to reconnect
-                    let _ = sender_to_fg.send_blocking(AsyncClientMessage::Connect);
-                }
-                _ => {}
-            }
-        }
-
-        for artist in res.into_iter() {
-            let _ = sender_to_fg.send_blocking(
-                AsyncClientMessage::RecentArtistDownloaded(artist),
-            );
-        }
-    }
-
-    pub fn fetch_songs_of_artist(
-        client: &mut mpd::Client<StreamWrapper>,
-        sender_to_fg: &Sender<AsyncClientMessage>,
-        name: String,
-    ) {
-        match fetch_songs_by_query(
-            client,
-            Query::new().and_with_op(
-                Term::Tag(Cow::Borrowed("artist")),
-                QueryOperation::Contains,
-                name.clone(),
-            ),
-            |songs| {
-                sender_to_fg.send_blocking(AsyncClientMessage::ArtistSongInfoDownloaded(
-                    name.clone(),
-                    songs,
-                ))
-            },
-        ) {
-            Err(true) => {
-                let _ = sender_to_fg.send_blocking(AsyncClientMessage::Connect);
-            }
-            _ => {}
-        }
-    }
-
-    pub fn fetch_folder_contents(
-        client: &mut mpd::Client<StreamWrapper>,
-        sender_to_fg: &Sender<AsyncClientMessage>,
-        path: String,
-    ) {
-        match client.lsinfo(&path) {
-            Ok(contents) => {
-                let _ = sender_to_fg
-                    .send_blocking(AsyncClientMessage::FolderContentsDownloaded(path, contents));
-            }
-            Err(MpdError::Io(_)) => {
-                // Connection error => attempt to reconnect
-                let _ = sender_to_fg.send_blocking(AsyncClientMessage::Connect);
-            }
-            _ => {}
-        }
-    }
-
-    pub fn fetch_playlist_songs(
-        client: &mut mpd::Client<StreamWrapper>,
-        sender_to_fg: &Sender<AsyncClientMessage>,
-        name: String,
-    ) {
-        if client.version.1 < 24 {
-            match client
-                .playlist(&name, Option::<Range<u32>>::None) {
-                    Ok(mut mpd_songs) => {
-                        let songs: Vec<SongInfo> = mpd_songs
-                            .iter_mut()
-                            .map(|mpd_song| SongInfo::from(std::mem::take(mpd_song)))
-                            .collect();
-                        if !songs.is_empty() {
-                            let _ = sender_to_fg.send_blocking(AsyncClientMessage::PlaylistSongInfoDownloaded(
-                                name.clone(),
-                                songs,
-                            ));
-                        }
-                    }
-                    Err(MpdError::Io(_)) => {
-                        // Connection error => attempt to reconnect
-                        let _ = sender_to_fg.send_blocking(AsyncClientMessage::Connect);
-                    }
-                    _ => {}
-                }
-        } else {
-            // For MPD 0.24+, use the new paged loading
-            let mut curr_len: u32 = 0;
-            let mut more: bool = true;
-            while more && (curr_len as usize) < FETCH_LIMIT {
-                match client
-                    .playlist(&name, Some(curr_len..(curr_len + BATCH_SIZE as u32))) {
-                        Ok(mut mpd_songs) => {
-                            let songs: Vec<SongInfo> = mpd_songs
-                                .iter_mut()
-                                .map(|mpd_song| SongInfo::from(std::mem::take(mpd_song)))
-                                .collect();
-                            more = songs.len() >= BATCH_SIZE as usize;
-                            if !songs.is_empty() {
-                                curr_len += songs.len() as u32;
-                                let _ = sender_to_fg.send_blocking(AsyncClientMessage::PlaylistSongInfoDownloaded(
-                                    name.clone(),
-                                    songs,
-                                ));
-                            }
-                        }
-                        Err(MpdError::Io(_)) => {
-                            // Connection error => attempt to reconnect
-                            let _ = sender_to_fg.send_blocking(AsyncClientMessage::Connect);
-                            return;
-                        }
-                        _ => {return;}
-                    }
-
-            }
-        }
-    }
-    pub fn fetch_songs_by_uri(client: &mut mpd::Client<StreamWrapper>, uris: &[&str]) -> Result<Vec<SongInfo>, bool> {
-        let mut res: Vec<mpd::Song> = Vec::with_capacity(uris.len());
-        for uri in uris.iter() {
-            match client.find(Query::new().and(Term::File, *uri), None) {
-                Ok(mut found_songs) => {
-                    if found_songs.len() > 0 {
-                        res.push(found_songs.remove(0));
-                    }
-                }
-                Err(MpdError::Io(_)) => {
-                    return Err(true);
-                }
-                _ => {
-                    return Err(false);
-                }
-            }
-        }
-
-        Ok(
-            res
-                .into_iter()
-                .map(|mut mpd_song| SongInfo::from(std::mem::take(&mut mpd_song)))
-                .collect()
-        )
-    }
-
-    pub fn fetch_last_n_songs(
-        client: &mut mpd::Client<StreamWrapper>,
-        sender_to_fg: &Sender<AsyncClientMessage>,
-        n: u32
-    ) {
-        let to_fetch: Vec<(String, OffsetDateTime)> = sqlite::get_last_n_songs(n).expect("Sqlite DB error");
-        match fetch_songs_by_uri(
-            client,
-            &to_fetch.iter().map(|tup| tup.0.as_str()).collect::<Vec<&str>>()
-        ) {
-            Ok(raw_songs) => {
-                let songs: Vec<SongInfo> = raw_songs
-                    .into_iter()
-                    .zip(
-                        to_fetch.iter().map(|r| r.1).collect::<Vec<OffsetDateTime>>()
-                    )
-                    .map(|mut tup| {
-                        tup.0.last_played = Some(tup.1);
-                        std::mem::take(&mut tup.0)
-                    })
-                    .collect();
-
-                if !songs.is_empty() {
-                    let _ = sender_to_fg.send_blocking(AsyncClientMessage::RecentSongInfoDownloaded(
-                        songs
-                    ));
-                }
-            }
-            Err(true) => {
-                // Connection error => attempt to reconnect
-                let _ = sender_to_fg.send_blocking(AsyncClientMessage::Connect);
-            }
-            _ => {}
-        }
-    }
-}
 
 #[derive(Debug)]
 pub struct MpdWrapper {
@@ -955,8 +246,8 @@ impl MpdWrapper {
                         BackgroundTask::FetchQueue => {
                             background::get_current_queue(&mut client, &sender_to_fg);
                         }
-                        BackgroundTask::FetchQueueChanges(version) => {
-                            background::get_queue_changes(&mut client, &sender_to_fg, version);
+                        BackgroundTask::FetchQueueChanges(version, total_len) => {
+                            background::get_queue_changes(&mut client, &sender_to_fg, version, total_len);
                         }
                         BackgroundTask::DownloadFolderCover(key) => {
                             background::download_folder_cover(
@@ -1662,7 +953,7 @@ impl MpdWrapper {
                                 if old_version == 0 {
                                     BackgroundTask::FetchQueue
                                 } else {
-                                    BackgroundTask::FetchQueueChanges(old_version)
+                                    BackgroundTask::FetchQueueChanges(old_version, status.queue_len)
                                 },
                                 true
                             );
