+++ src/library/album_content_view.rs
@@ -403,6 +403,8 @@ impl AlbumContentView {
                         wiki_link.set_visible(false);
                     }
                     wiki_attrib.set_label(&wiki.attribution);
+                } else {
+                    wiki_text.set_label("");
                 }
                 if infobox_spinner.visible_child_name().unwrap() != "content" {
                     infobox_spinner.set_visible_child_name("content");
